+++ crates/unftp-auth-jsonfile/src/lib.rs
@@ -78,6 +78,8 @@
 
 use async_trait::async_trait;
 use bytes::Bytes;
+use ipnet::Ipv4Net;
+use iprange::IpRange;
 use libunftp::auth::{AuthenticationError, Authenticator, DefaultUser};
 use ring::{
     digest::SHA256_OUTPUT_LEN,
@@ -94,21 +96,21 @@ enum Credentials {
     Plaintext {
         username: String,
         password: String,
+        ip_ranges: Option<Vec<String>>,
     },
     Pbkdf2 {
         username: String,
         pbkdf2_salt: String,
         pbkdf2_key: String,
         pbkdf2_iter: NonZeroU32,
+        ip_ranges: Option<Vec<String>>,
     },
 }
 
 /// This structure implements the libunftp `Authenticator` trait
-///
-
 #[derive(Clone, Debug)]
 pub struct JsonFileAuthenticator {
-    credentials_map: HashMap<String, Password>,
+    credentials_map: HashMap<String, UserCreds>,
 }
 
 #[derive(Clone, Debug)]
@@ -123,51 +125,78 @@ enum Password {
     },
 }
 
+#[derive(Clone, Debug)]
+struct UserCreds {
+    pub password: Password,
+    pub ip_ranges: Option<IpRange<Ipv4Net>>,
+}
+
 impl JsonFileAuthenticator {
     /// Initialize a new [`JsonFileAuthenticator`] from file.
     pub fn from_file<P: AsRef<Path>>(filename: P) -> Result<Self, Box<dyn std::error::Error>> {
         let json: String = fs::read_to_string(filename)?;
-
-        JsonFileAuthenticator::from_json(json)
+        Self::from_json(json)
     }
 
     /// Initialize a new [`JsonFileAuthenticator`] from json string.
     pub fn from_json<T: Into<String>>(json: T) -> Result<Self, Box<dyn std::error::Error>> {
         let credentials_list: Vec<Credentials> = serde_json::from_str::<Vec<Credentials>>(&json.into())?;
-        let map: Result<HashMap<String, Password>, _> = credentials_list.into_iter().map(Self::list_entry_to_map_entry).collect();
+        let map: Result<HashMap<String, UserCreds>, _> = credentials_list.into_iter().map(Self::list_entry_to_map_entry).collect();
         Ok(JsonFileAuthenticator { credentials_map: map? })
     }
 
-    fn list_entry_to_map_entry(user_info: Credentials) -> Result<(String, Password), Box<dyn std::error::Error>> {
+    fn list_entry_to_map_entry(user_info: Credentials) -> Result<(String, UserCreds), Box<dyn std::error::Error>> {
         let map_entry = match user_info {
-            Credentials::Plaintext { username, password } => (username, Password::PlainPassword { password }),
+            Credentials::Plaintext { username, password, ip_ranges } => (
+                username.clone(),
+                UserCreds {
+                    password: Password::PlainPassword { password },
+                    ip_ranges: Self::parse_ip_range(username, ip_ranges)?,
+                },
+            ),
             Credentials::Pbkdf2 {
                 username,
                 pbkdf2_salt,
                 pbkdf2_key,
                 pbkdf2_iter,
+                ip_ranges,
             } => (
                 username.clone(),
-                Password::Pbkdf2Password {
-                    pbkdf2_salt: base64::decode(pbkdf2_salt)
-                        .map_err(|_| "Could not base64 decode the salt")?
-                        .try_into()
-                        .map_err(|_| "Could not convert String to Bytes")?,
-                    pbkdf2_key: base64::decode(pbkdf2_key)
-                        .map_err(|_| "Could not decode base64")?
-                        .validate("pbkdf2_key", &Length::Max(SHA256_OUTPUT_LEN))
-                        .result()
-                        .map_err(|_| {
-                            format!("Key of user \"{}\" is too long", username)
-                        })?
-                        .unwrap() // this unwrap is just giving the value within
-                        .try_into()
-                        .map_err(|_| "Could not convert to Bytes")?,
-                    pbkdf2_iter,
+                UserCreds {
+                    password: Password::Pbkdf2Password {
+                        pbkdf2_salt: base64::decode(pbkdf2_salt)
+                            .map_err(|_| "Could not base64 decode the salt")?
+                            .try_into()
+                            .map_err(|_| "Could not convert String to Bytes")?,
+                        pbkdf2_key: base64::decode(pbkdf2_key)
+                            .map_err(|_| "Could not decode base64")?
+                            .validate("pbkdf2_key", &Length::Max(SHA256_OUTPUT_LEN))
+                            .result()
+                            .map_err(|_| {
+                                format!("Key of user \"{}\" is too long", username)
+                            })?
+                            .unwrap() // this unwrap is just giving the value within
+                            .try_into()
+                            .map_err(|_| "Could not convert to Bytes")?,
+                        pbkdf2_iter,
+                    },
+                    ip_ranges: Self::parse_ip_range(username, ip_ranges)?,
                 },
             ),
         };
         Ok(map_entry)
+    }
+
+    fn parse_ip_range(username: String, ip_ranges: Option<Vec<String>>) -> Result<Option<IpRange<Ipv4Net>>, String> {
+        ip_ranges
+            .map(|v| {
+                let range: Result<IpRange<Ipv4Net>, _> = v
+                    .iter()
+                    .map(|s| s.parse::<Ipv4Net>().map_err(|_| format!("could not parse IP ranges for user {}", username)))
+                    .collect();
+                range
+            })
+            .transpose()
     }
 
     fn check_password(given_password: &str, actual_password: &Password) -> Result<(), ()> {
@@ -186,23 +215,42 @@ impl JsonFileAuthenticator {
             } => verify(PBKDF2_HMAC_SHA256, *pbkdf2_iter, pbkdf2_salt, given_password.as_bytes(), pbkdf2_key).map_err(|_| ()),
         }
     }
+
+    fn ip_ok(creds: &libunftp::auth::Credentials, actual_creds: &UserCreds) -> bool {
+        match &actual_creds.ip_ranges {
+            Some(allowed) => match creds.source_ip {
+                std::net::IpAddr::V4(ref ip) => allowed.contains(ip),
+                _ => false,
+            },
+            None => true,
+        }
+    }
 }
 
 #[async_trait]
 impl Authenticator<DefaultUser> for JsonFileAuthenticator {
     #[tracing_attributes::instrument]
-    async fn authenticate(&self, username: &str, given_password: &str) -> Result<DefaultUser, AuthenticationError> {
-        if let Some(actual_password) = self.credentials_map.get(username) {
-            if let Ok(()) = Self::check_password(given_password, actual_password) {
-                return Ok(DefaultUser);
-            } else {
-                sleep(Duration::from_millis(1500)).await;
-                return Err(AuthenticationError::BadPassword);
+    async fn authenticate(&self, username: &str, creds: &libunftp::auth::Credentials) -> Result<DefaultUser, AuthenticationError> {
+        let res = if let Some(actual_creds) = self.credentials_map.get(username) {
+            match creds.password {
+                None => Err(AuthenticationError::BadPassword),
+                Some(ref given_password) => {
+                    if Self::check_password(given_password, &actual_creds.password).is_ok() {
+                        Self::ip_ok(creds, &actual_creds)
+                            .then(|| Ok(DefaultUser {}))
+                            .unwrap_or(Err(AuthenticationError::IpDisallowed))
+                    } else {
+                        Err(AuthenticationError::BadPassword)
+                    }
+                }
             }
         } else {
-            sleep(Duration::from_millis(1500)).await;
             Err(AuthenticationError::BadUser)
+        };
+        if res.is_err() {
+            sleep(Duration::from_millis(1500)).await;
         }
+        res
     }
 }
 
@@ -230,35 +278,36 @@ mod test {
   },
   {
     "username": "dan",
-    "password": ""
+    "password": "",
+    "ip_ranges": ["127.0.0.1/24"]
   }
 ]"#;
         let json_authenticator = JsonFileAuthenticator::from_json(json).unwrap();
         assert_eq!(
             json_authenticator
-                .authenticate("alice", "this is the correct password for alice")
+                .authenticate("alice", &"this is the correct password for alice".into())
                 .await
                 .unwrap(),
             DefaultUser
         );
         assert_eq!(
             json_authenticator
-                .authenticate("bella", "this is the correct password for bella")
+                .authenticate("bella", &"this is the correct password for bella".into())
                 .await
                 .unwrap(),
             DefaultUser
         );
-        assert_eq!(json_authenticator.authenticate("carol", "not so secure").await.unwrap(), DefaultUser);
-        assert_eq!(json_authenticator.authenticate("dan", "").await.unwrap(), DefaultUser);
-        match json_authenticator.authenticate("carol", "this is the wrong password").await {
+        assert_eq!(json_authenticator.authenticate("carol", &"not so secure".into()).await.unwrap(), DefaultUser);
+        assert_eq!(json_authenticator.authenticate("dan", &"".into()).await.unwrap(), DefaultUser);
+        match json_authenticator.authenticate("carol", &"this is the wrong password".into()).await {
             Err(AuthenticationError::BadPassword) => assert!(true),
             _ => assert!(false),
         }
-        match json_authenticator.authenticate("bella", "this is the wrong password").await {
+        match json_authenticator.authenticate("bella", &"this is the wrong password".into()).await {
             Err(AuthenticationError::BadPassword) => assert!(true),
             _ => assert!(false),
         }
-        match json_authenticator.authenticate("chuck", "12345678").await {
+        match json_authenticator.authenticate("chuck", &"12345678".into()).await {
             Err(AuthenticationError::BadUser) => assert!(true),
             _ => assert!(false),
         }