+++ crates/unftp-auth-pam/src/lib.rs
@@ -29,14 +29,14 @@ impl PamAuthenticator {
 impl Authenticator<DefaultUser> for PamAuthenticator {
     #[allow(clippy::type_complexity)]
     #[tracing_attributes::instrument]
-    async fn authenticate(&self, username: &str, password: &str) -> Result<DefaultUser, AuthenticationError> {
-        let service = self.service.clone();
+    async fn authenticate(&self, username: &str, creds: &Credentials) -> Result<DefaultUser, AuthenticationError> {
         let username = username.to_string();
-        let password = password.to_string();
+        let password = creds.password.as_ref().ok_or(AuthenticationError::BadPassword)?;
+        let service = self.service.clone();
 
         let mut auth = pam_auth::Authenticator::with_password(&service).map_err(|e| AuthenticationError::with_source("pam error", e))?;
 
-        auth.get_handler().set_credentials(&username, &password);
+        auth.get_handler().set_credentials(&username, password);
         auth.authenticate().map_err(|e| AuthenticationError::with_source("pam error", e))?;
         Ok(DefaultUser {})
     }